fix(mcp): merge curator listing routes into [slug]/listing
Next.js forbids two different param names ([id] and [slug]) at the same
dynamic path level. The Phase 41 executor split mutations under [id]/listing
and reads under [slug]/listing, which collided with the existing [slug]/call
route and broke pnpm dev with "You cannot use different slug names for the
same dynamic path".

Merge POST/PATCH/DELETE handlers into the existing [slug]/listing route file.
The URL segment is named slug at the route level (consistent with [slug]/call)
but mutation handlers internally rename it to workflowId since that is what
the segment carries for those methods.

Update unit-test fixtures (mcp-curator-tools, mcp-curator-slug,
mcp-curator-price-change) to use the new import path and the slug param key.

Caught during local smoke; all curator unit tests still pass.
